pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.0.1"
description = "Template project to use as base for my python projects"
authors = ["Ítalo Da Silva Vieira <italo.vieira.dev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.taskipy.tasks]
install = "poetry install"
post_install = "pre-commit install"
start = "python src/main.py"
test = "pytest"
bandit = "bandit -r src"
black = "black src"
mypy = "mypy --strict src"
ruff = "ruff src"
pre_commit = "task black && task bandit && task mypy && task ruff"
export = "poetry export -f requirements.txt --output requirements.txt"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
pandas = "^2.0.3"
lxml = "^4.9.3"
html5lib = "^1.1"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.11.0"
black = "^23.3.0"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
bandit = "^1.7.5"
ruff = "^0.0.270"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"